requirements.txt
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
dotenv==0.9.9
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
Telethon==1.42.0
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
